move: enum폴더 위치 변경
