+++ src/main/java/step3/domain/LottoManager.java
@@ -4,14 +4,14 @@ import static java.util.Collections.shuffle;
 import static step3.constant.LottoInfoConstant.LOTTO_ELEMENTS_SIZE;
 import static step3.constant.LottoInfoConstant.LOTTO_PRICE;
 import static step3.constant.LottoInfoConstant.LOTTO_VALID_ELEMENTS;
-import static step3.constant.enums.LottoReward.numberToLottoReward;
+import static step3.enums.LottoReward.numberToLottoReward;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.stream.Collectors;
-import step3.constant.enums.LottoReward;
+import step3.enums.LottoReward;
 
 public class LottoManager {
